requirements.txt
fastapi==0.115.12
jinja2==3.1.6
pydantic-settings==2.8.1
PyYAML==6.0.2
requests==2.32.3
uvicorn==0.34.0
